% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/onLoad.R
\name{optifunset}
\alias{optifunset}
\title{OptIfUnset Package}
\description{
Set an option only if is not currently set
}
\details{
This package contains a single function which only updates the global options if the option is presently unset
}
\examples{
options.ifunset(width=100)  #NO CHANGE, ALREADY EXISTS
options.ifunset(myuniqueoption=TRUE) #NEW Option Created
}

