% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/options-ifunset.R
\name{options.ifunset}
\alias{options.ifunset}
\title{Set Options if Unset}
\usage{
options.ifunset(..., force = FALSE)
}
\arguments{
\item{...}{any options can be defined, using \code{name = value}, if they are not already defined.

Options can also be passed by giving a single unnamed argument which is a named list.}

\item{force}{Force the Option to Be Set}
}
\description{
This function will set an option if it isn't already present within the global options
returned by the \code{options()} function
}
\examples{
options.ifunset(width=10)            #IGNORED, ALREADY EXISTS
options.ifunset(width=10,force=TRUE) #FORCED UPDATE TO OPTION
options.ifunset(myoption=TRUE)       #New Option is Created
}

